+++ test/convert.test.js
@@ -1,23 +1,23 @@
 // Import necessary modules
-import { execSync } from 'child_process';
-import { expect } from 'chai';
-import fs from 'fs';
-
+import { execSync } from "child_process";
+import { expect } from "chai";
+import fs from "fs";
 
 // Describe block for 'convert' command tests
-describe('Convert Command Tests', function () {
-  
+describe("Convert Command Tests", function () {
   // Test case 1: Proper command
-  it('should convert image from PNG to JPEG', function () {
-    const result = execSync('node ./index.js convert ./myimg.png ./myimg.jpeg').toString();
-    expect(result).to.include('Image converted from .png to .jpeg');
+  it("should convert image from PNG to JPEG", function () {
+    const result = execSync(
+      "node ./index.js convert ./myimg.png ./myimg.jpeg"
+    ).toString();
+    expect(result).to.include("Image converted from .png to .jpeg");
   });
 
   // Test case 2: When outputFile arg not given
-  it('should handle missing output file argument', function () {
+  it("should handle missing output file argument", function () {
     try {
       // Run the command and capture the result
-      const result = execSync('node ./index.js convert ./myimg.png');
+      const result = execSync("node ./index.js convert ./myimg.png");
     } catch (error) {
       // Check if the expected error message is present in the stderr
       const errMsg = "error: missing required argument 'outputFile'";
@@ -25,10 +25,10 @@ describe('Convert Command Tests', function () {
     }
   });
   // // Test case 3: When both the args outputFile and inputFile are missing
-  it('should handle missing input and output file arguments', function () {
+  it("should handle missing input and output file arguments", function () {
     try {
       // Run the command and capture the result
-      const result = execSync('node ./index.js convert');
+      const result = execSync("node ./index.js convert");
     } catch (error) {
       // Check if the expected error message is present in the stderr
       const errMsg = "error: missing required argument 'inputFile'";
@@ -37,10 +37,12 @@ describe('Convert Command Tests', function () {
   });
 
   // // Test case 4: When inputFile doesn't exist
-  it('should handle non-existent inputFile argument', function () {
+  it("should handle non-existent inputFile argument", function () {
     try {
       // Run the command and capture the result
-      const result = execSync('node ./index.js convert ./myimg.jpg ./myimg.tiff');
+      const result = execSync(
+        "node ./index.js convert ./myimg.jpg ./myimg.tiff"
+      );
     } catch (error) {
       // Check if the expected error message is present in the stderr
       const errMsg = "Input file is missing: ./myimg.jpg";
@@ -51,6 +53,8 @@ describe('Convert Command Tests', function () {
   // Cleanup after testing
   after(function () {
     // Delete temporary file created during testing
-    fs.unlinkSync('./myimg.jpeg');
+    if (fs.existsSync("./myimg.jpeg")) {
+      fs.unlinkSync("./myimg.jpeg");
+    }
   });
 });
